requirements.txt
kafka-python
psycopg2
jsonschema
streamlit
plotly
pydeck
flask
requests
